(#001) Initialization of the Content of ATM CV Project
This initial commit generates all content for creation of the
project, alongside this aggregation of various parts that comprise
the ATM CV Template. This commit serves as the most crucial piece for
the project's formation, laying the groundwork for a primary aggregation
of essential information. With this, fundamental elements are established
that will be developed and enriched in future updates, ensuring coherence
and structure from the very beginning.

An initial creation has been integrated, including the GNU licence and the
Security Policy. Furthermore, the main template has been designed using
Typst, ensuring a professional format and style from the outset. A TODO
 file in Markdown has also been generated for tracking pending tasks, along
with Bundler documentation, facilitating dependency mangament and continuous
development.

Signed-off-by: Joseph Gallegos <jaremgallegosp@gmail.com>